src/test/java/students/arturs_arutjunovs/lesson_14/level_1/task_1_to_6/FruitStorageTest.java
package students.arturs_arutjunovs.lesson_14.level_1.task_1_to_6;

import org.junit.jupiter.api.Test;

import java.util.ArrayList;
import java.util.List;

import static org.junit.jupiter.api.Assertions.assertEquals;

class FruitStorageTest {
    FruitStorage fruitStorage = new FruitStorage();

    @Test
    public void findRedApples() {
        List<Apple> inventory = fruitStorage.getAllApples();
        List<Apple> expected = new ArrayList<>();
        expected.add(new Apple("red", 100));
        expected.add(new Apple("red", 160));
        expected.add(new Apple("red", 200));
        assertEquals(expected, fruitStorage.filterApplesByColor(inventory, "red"));
    }

    @Test
    public void findGreenApples() {
        List<Apple> inventory = fruitStorage.getAllApples();
        List<Apple> expected = new ArrayList<>();
        expected.add(new Apple("green", 200));
        expected.add(new Apple("green", 90));
        expected.add(new Apple("green", 50));
        assertEquals(expected, fruitStorage.filterApplesByColor(inventory, "green"));
    }

    @Test
    public void findHeavyApples() {  //more than 150 gr.
        List<Apple> inventory = fruitStorage.getAllApples();
        List<Apple> expected = new ArrayList<>();
        expected.add(new Apple("red", 160));
        expected.add(new Apple("red", 200));
        expected.add(new Apple("green", 200));
        expected.add(new Apple("yellow", 170));
        assertEquals(expected, fruitStorage.filterHeavyApples(inventory, 150));
    }

    @Test
    public void findLightApples() {  //less than 150 gr.
        List<Apple> inventory = fruitStorage.getAllApples();
        List<Apple> expected = new ArrayList<>();
        expected.add(new Apple("red", 100));
        expected.add(new Apple("green", 90));
        expected.add(new Apple("green", 50));
        expected.add(new Apple("yellow", 50));
        assertEquals(expected, fruitStorage.filterLightApples(inventory, 150));
    }
}